Remove flag-icon-css
I originally stopped using it because I need to know
the country-code but only had the language-code
at that time.

By now I decided against showing a flag in the language select.
The issue already arises when choosing the flag for english. USA, Great Britain, England, ...? Which one is best?